small github refactor
